Ugly attempt to simulate vbridge and test in-GPU copying
New step in improvement is to get from vmem to vbridge, this vlc's
vout module don't do any GPU-CPU copy and pass directly in-GPU datas.
But before doing this, I wanted to test in-GPU copy from one texture
to another.

This code simulate vbridge by getting pixels data and stuffing them into
a texture. But there is a problem I discovered : in libvlc's side, we
don't have any OpenGL context defined, so we can't do that, it doesn't
seem to work. I'll have to solve this problem later, now I juste
initialized the texture in Unity's callback where I have an OGL
context defined.
One clue I have to follow is to create a context sharing data with
Unity's one.

The other part in the code is a successfull in-GPU copy.

For that I have to declare a new FBO (during libvlc init) that I need
to bound to in Unity's cb before calling CopyTexSubImage2D that
perform in-GPU copy of previously read data to unity's texture

This code is dirty and still full of debug