d3d11: fix leaks by reworking how resources are allocated and deallocated
Unity provides some resources on startup (unless it fails to initialize d3d11 properly).
We create a d3d11 device.
When the cleanup callback is called, we can release most of the resources but not our d3d11 device.
When the mediaplayer is released, we can unsub from the callbacks and release the d3d11 device we created.