Editor: Free VLCUnityPlugin handle when stopping scene in Editor
Unity loads the VLCUnity native plugin but does not unload it. This is a known Unity issue discussed online many times.
The issue it causes is that the VLCUnityPlugin.dll is not unloaded by the Editor which prevents developers to replace or move it,
when troubleshooting in Editor builds.

Close https://code.videolan.org/videolan/vlc-unity/-/issues/98