Editor: Use AssetPostprocessor for libvlc configuration
the main point of this to speed up dramatically importing the asset onto the editor. All platforms are covered.
There are 2 Unity quirks that are worked around, one using an additional preprocess function for VLCUnityPlugin to avoid an incorrect error, the other is for macos, where we clear the console from incorrect errors (that disappear on build otherwise) to avoid confusing the user.